+++ pkg/api/api.go
@@ -57,7 +57,7 @@ type Relay interface {
 	GetBlockReceived(context.Context, structs.HeaderTraceQuery) ([]structs.BidTraceWithTimestamp, error)
 }
 
-type Registrations interface {
+type Validators interface {
 	// Proposer APIs (builder spec https://github.com/ethereum/builder-specs)
 	RegisterValidator(context.Context, *structs.MetricGroup, []types.SignedValidatorRegistration) error
 	// Data APIs
@@ -69,13 +69,13 @@ type Registrations interface {
 type API struct {
 	l   log.Logger
 	r   Relay
-	reg Registrations
+	val Validators
 
 	m APIMetrics
 }
 
-func NewApi(l log.Logger, r Relay, reg Registrations) (a *API) {
-	a = &API{l: l, r: r, reg: reg}
+func NewApi(l log.Logger, r Relay, val Validators) (a *API) {
+	a = &API{l: l, r: r, val: val}
 	a.initMetrics()
 	return a
 }
@@ -145,7 +145,7 @@ func (a *API) registerValidator(w http.ResponseWriter, r *http.Request) (status 
 	}
 
 	m := structs.NewMetricGroup(4)
-	if err = a.reg.RegisterValidator(r.Context(), m, payload); err != nil {
+	if err = a.val.RegisterValidator(r.Context(), m, payload); err != nil {
 		m.ObserveWithError(a.m.RelayTiming, err)
 		a.m.ApiReqCounter.WithLabelValues("registerValidator", "400", "register validator").Inc()
 		a.l.With(log.F{
@@ -279,7 +279,7 @@ func (a *API) getValidators(w http.ResponseWriter, r *http.Request) (int, error)
 	defer timer.ObserveDuration()
 
 	m := structs.NewMetricGroup(4)
-	vs := a.reg.GetValidators(m)
+	vs := a.val.GetValidators(m)
 	if vs == nil {
 		a.l.Trace("no registered validators for epoch")
 		vs = structs.BuilderGetValidatorsResponseEntrySlice{}
@@ -313,7 +313,7 @@ func (a *API) specificRegistration(w http.ResponseWriter, r *http.Request) (int,
 		return http.StatusBadRequest, err
 	}
 
-	registration, err := a.reg.Registration(r.Context(), pk)
+	registration, err := a.val.Registration(r.Context(), pk)
 	if err != nil {
 		a.m.ApiReqCounter.WithLabelValues("specificRegistration", "500", "registration").Inc()
 		return http.StatusInternalServerError, err
