+++ pkg/datastore/block/dbadger/datastore.go
@@ -7,13 +7,14 @@ import (
 	"sync"
 	"time"
 
+	"github.com/blocknative/dreamboat/pkg/datastore/block/headerscontroller"
 	"github.com/blocknative/dreamboat/pkg/structs"
-	"github.com/dgraph-io/badger/v2"
 	"github.com/flashbots/go-boost-utils/types"
 	lru "github.com/hashicorp/golang-lru/v2"
 	ds "github.com/ipfs/go-datastore"
 )
 
+/*
 type TTLStorage interface {
 	PutWithTTL(context.Context, ds.Key, []byte, time.Duration) error
 	Get(context.Context, ds.Key) ([]byte, error)
@@ -25,26 +26,30 @@ type Badger interface {
 	View(func(txn *badger.Txn) error) error
 	Update(func(txn *badger.Txn) error) error
 	NewTransaction(bool) *badger.Txn
+}*/
+
+type DB interface {
+	PutWithTTL(context.Context, ds.Key, []byte, time.Duration) error
+	Get(context.Context, ds.Key) ([]byte, error)
+	GetBatch(ctx context.Context, keys []ds.Key) (batch [][]byte, err error)
 }
 
 type Datastore struct {
-	TTLStorage
-	Badger
+	DB
 	PayloadCache *lru.Cache[structs.PayloadKey, *structs.BlockBidAndTrace]
 
-	hc *HeaderController
+	hc *headerscontroller.HeaderController
 	l  sync.Mutex
 }
 
-func NewDatastore(t TTLStorage, v Badger, hc *HeaderController, payloadCacheSize int) (*Datastore, error) {
+func NewDatastore(t DB, hc *headerscontroller.HeaderController, payloadCacheSize int) (*Datastore, error) {
 	cache, err := lru.New[structs.PayloadKey, *structs.BlockBidAndTrace](payloadCacheSize)
 	if err != nil {
 		return nil, err
 	}
 
 	return &Datastore{
-		TTLStorage:   t,
-		Badger:       v,
+		DB:           t,
 		hc:           hc,
 		PayloadCache: cache,
 	}, nil
@@ -56,93 +61,12 @@ func (s *Datastore) CacheBlock(ctx context.Context, block *structs.CompleteBlock
 	return nil
 }
 
-func (s *Datastore) PutDelivered(ctx context.Context, slot structs.Slot, trace structs.DeliveredTrace, ttl time.Duration) error {
-	data, err := json.Marshal(trace.Trace)
-	if err != nil {
-		return err
-	}
-
-	txn := s.Badger.NewTransaction(true)
-	defer txn.Discard()
-	if err := txn.SetEntry(badger.NewEntry(DeliveredHashKey(trace.Trace.BlockHash).Bytes(), DeliveredKey(slot).Bytes()).WithTTL(ttl)); err != nil {
-		return err
-	}
-	if err := txn.SetEntry(badger.NewEntry(DeliveredNumKey(trace.BlockNumber).Bytes(), DeliveredKey(slot).Bytes()).WithTTL(ttl)); err != nil {
-		return err
-	}
-	if err := txn.SetEntry(badger.NewEntry(DeliveredPubkeyKey(trace.Trace.ProposerPubkey).Bytes(), DeliveredKey(slot).Bytes()).WithTTL(ttl)); err != nil {
-		return err
-	}
-	if err := txn.SetEntry(badger.NewEntry(DeliveredKey(slot).Bytes(), data).WithTTL(ttl)); err != nil {
-		return err
-	}
-
-	return txn.Commit()
-}
-
-func (s *Datastore) CheckSlotDelivered(ctx context.Context, slot uint64) (bool, error) {
-	tx := s.Badger.NewTransaction(false)
-	defer tx.Discard()
-
-	_, err := tx.Get(DeliveredKey(structs.Slot(slot)).Bytes())
-	if err == badger.ErrKeyNotFound {
-		return false, nil
-	}
-	return (err == nil), err
-}
-
-func (s *Datastore) GetDelivered(ctx context.Context, query structs.PayloadQuery) (structs.BidTraceWithTimestamp, error) {
-	key, err := s.queryToDeliveredKey(ctx, query)
-	if err != nil {
-		return structs.BidTraceWithTimestamp{}, err
-	}
-	return s.getDelivered(ctx, key)
-}
-
-func (s *Datastore) getDelivered(ctx context.Context, key ds.Key) (structs.BidTraceWithTimestamp, error) {
-	data, err := s.TTLStorage.Get(ctx, key)
-	if err != nil {
-		return structs.BidTraceWithTimestamp{}, err
-	}
-
-	var trace structs.BidTraceWithTimestamp
-	err = json.Unmarshal(data, &trace)
-	return trace, err
-}
-
-func (s *Datastore) GetDeliveredBatch(ctx context.Context, queries []structs.PayloadQuery) ([]structs.BidTraceWithTimestamp, error) {
-	keys := make([]ds.Key, 0, len(queries))
-	for _, query := range queries {
-		key, err := s.queryToDeliveredKey(ctx, query)
-		if err != nil {
-			return nil, err
-		}
-		keys = append(keys, key)
-	}
-
-	batch, err := s.TTLStorage.GetBatch(ctx, keys)
-	if err != nil {
-		return nil, err
-	}
-
-	traceBatch := make([]structs.BidTraceWithTimestamp, 0, len(batch))
-	for _, data := range batch {
-		var trace structs.BidTraceWithTimestamp
-		if err = json.Unmarshal(data, &trace); err != nil {
-			return nil, err
-		}
-		traceBatch = append(traceBatch, trace)
-	}
-
-	return traceBatch, err
-}
-
 func (s *Datastore) PutPayload(ctx context.Context, key structs.PayloadKey, payload *structs.BlockBidAndTrace, ttl time.Duration) error {
 	data, err := json.Marshal(payload)
 	if err != nil {
 		return err
 	}
-	return s.TTLStorage.PutWithTTL(ctx, PayloadKeyKey(key), data, ttl)
+	return s.DB.PutWithTTL(ctx, PayloadKeyKey(key), data, ttl)
 }
 
 func (s *Datastore) GetPayload(ctx context.Context, key structs.PayloadKey) (*structs.BlockBidAndTrace, bool, error) {
@@ -151,35 +75,13 @@ func (s *Datastore) GetPayload(ctx context.Context, key structs.PayloadKey) (*st
 		return memPayload, true, nil
 	}
 
-	data, err := s.TTLStorage.Get(ctx, PayloadKeyKey(key))
+	data, err := s.DB.Get(ctx, PayloadKeyKey(key))
 	if err != nil {
 		return nil, false, err
 	}
 	var payload structs.BlockBidAndTrace
 	err = json.Unmarshal(data, &payload)
 	return &payload, false, err
-}
-
-func (s *Datastore) queryToDeliveredKey(ctx context.Context, query structs.PayloadQuery) (ds.Key, error) {
-	var (
-		rawKey []byte
-		err    error
-	)
-
-	if (query.BlockHash != types.Hash{}) {
-		rawKey, err = s.TTLStorage.Get(ctx, DeliveredHashKey(query.BlockHash))
-	} else if query.BlockNum != 0 {
-		rawKey, err = s.TTLStorage.Get(ctx, DeliveredNumKey(query.BlockNum))
-	} else if (query.PubKey != types.PublicKey{}) {
-		rawKey, err = s.TTLStorage.Get(ctx, DeliveredPubkeyKey(query.PubKey))
-	} else {
-		rawKey = DeliveredKey(query.Slot).Bytes()
-	}
-
-	if err != nil {
-		return ds.Key{}, err
-	}
-	return ds.NewKey(string(rawKey)), nil
 }
 
 func HeaderHashKey(bh types.Hash) ds.Key {
@@ -190,42 +92,11 @@ func HeaderNumKey(bn uint64) ds.Key {
 	return ds.NewKey(fmt.Sprintf("header-num-%d", bn))
 }
 
-func DeliveredKey(slot structs.Slot) ds.Key {
-	return ds.NewKey(fmt.Sprintf("delivered-%d", slot))
-}
-
-func DeliveredHashKey(bh types.Hash) ds.Key {
-	return ds.NewKey(fmt.Sprintf("delivered-hash-%s", bh.String()))
-}
-
-func DeliveredNumKey(bn uint64) ds.Key {
-	return ds.NewKey(fmt.Sprintf("delivered-num-%d", bn))
-}
-
-func DeliveredPubkeyKey(pk types.PublicKey) ds.Key {
-	return ds.NewKey(fmt.Sprintf("delivered-pk-%s", pk.String()))
-}
-
 func PayloadKeyKey(key structs.PayloadKey) ds.Key {
 	return ds.NewKey(fmt.Sprintf("payload-%s-%s-%d", key.BlockHash.String(), key.Proposer.String(), key.Slot))
 }
 
+/*
 func ValidatorKey(pk structs.PubKey) ds.Key {
 	return ds.NewKey(fmt.Sprintf("valdator-%s", pk.String()))
-}
-
-type TTLDatastoreBatcher struct {
-	ds.TTLDatastore
-}
-
-func (bb *TTLDatastoreBatcher) GetBatch(ctx context.Context, keys []ds.Key) (batch [][]byte, err error) {
-	for _, key := range keys {
-		data, err := bb.TTLDatastore.Get(ctx, key)
-		if err != nil {
-			continue
-		}
-		batch = append(batch, data)
-	}
-
-	return
-}
+}*/
